phase 1 refactor - moving stuff around
