+++ Dockerfile
@@ -20,12 +20,8 @@ FROM python:3.13-slim
 
 WORKDIR /foul-play
 
-COPY config.py /foul-play/config.py
-COPY constants.py /foul-play/constants.py
-COPY data /foul-play/data
 COPY run.py /foul-play/run.py
 COPY fp /foul-play/fp
-COPY teams /foul-play/teams
 
 COPY --from=build /packages/ /usr/local/lib/python3.13/site-packages/
