+++ fp/data/mods/gen1_pokedex_mods.json
