+++ fp/data/mods/gen6_move_mods.json
