+++ fp/data/mods/gen6_pokedex_mods.json
