+++ fp/data/mods/gen8_pokedex_mods.json
